Warning Squashing

+++ dev/Model/state/state.h
@@ -32,14 +32,14 @@ struct device_state
 	uint16_t pack()
 	{
 		uint16_t packed = switches_pack() << 8;
-		unsigned char pos = position * 255;
+		unsigned char pos = static_cast<int>(position) * 255;
 		packed += pos;
 
 		return packed;
 	}
 	void unpack(uint16_t state)
 	{
-		position = (state & 0xff)/255;
+		position = (state & 0xff)/255.0f;
 		switches_unpack(state >> 8);
 	}
